--- a/src/plugins/vis_augmenter/public/saved_augment_vis/saved_augment_vis.ts
+++ b/src/plugins/vis_augmenter/public/saved_augment_vis/saved_augment_vis.ts
@@ -89,6 +89,8 @@ export class SavedObjectLoaderAugmentVis extends SavedObjectLoader {
    * @param search
    * @param size
    * @param fields
+   * @param hasReference Optional field to specify a reference
+   * @param searchFields Optional field to specify the search fields in the query
    * @returns {Promise}
    */
   findAll(