### Description

This PR provides optimizations to the saved object loader calls made in the vis embeddable. Specifically, when fetching the saved objects, it adds arguments to the query to filter by matching plugin resource IDs, when applicable. Before, all objects were returned and post-processing filtering was done. Now, we leverage OpenSearch to handle the filtering for us.

To make this change, a new optional argument has been added to the `findAll()` function to specify the search fields. Before, it was hardcoded to name and description. Now, we want to search on the `pluginResource.id` field such that we only return objects that match the input resource IDs.

This fix provides massive performance improvements when rendering the View Events flyout. Specifically, it prevents `n^2 - n` objects being returned, where `n` is the number of associations. This also means an equivalent amount of prevented plugin expression function runs, any queries ran within those expression functions, etc.

For example, suppose there are 10 associations to a visualization, and a user opens the View Events flyout.
Before, each visualization would be fetching all 10 `augment-vis` objects, executing all 10 expression functions (which would mean fetching all 10 resources, results for each of them, etc.). Each vis (10 executions) x 10 == 100 executions.
Now, each visualization only fetches the relevant `augment-vis` saved object, so only one expression function is ran. Each vis (1 execution) x 10 = 10 executions, a savings of 90%.

Before this change, the time to render the view events flyout as the number of visualizations/associations grew was exponential. Now, it is linear. Example of network request savings seen on my local:
Opening flyout with 5 associations = 98 requests (before) - 58 requests (after) = 40 requests prevented = ~41% less total requests.
10 associations = 288 requests (before) - 108 requests (after) = 180 requests prevented = ~63% less total requests

For more information on plugin resource IDs and how they are used, see "The VisAugmenterEmbeddableConfig" section in #3415 

Quick results on my local (top row is association count)
```
|        | 10   | 20   | 40   | 50   |
|--------|------|------|------|------|
| before | 1.7s | 3.4s | 5.4s | 9.3s |
| after  | 1.3s | 2.0s | 2.5s | 2.9s |
```

### Check List

- [ ] All tests pass
  - [ ] `yarn test:jest`
  - [ ] `yarn test:jest_integration`
  - [ ] `yarn test:ftr`
- [ ] New functionality includes testing.
- [ ] New functionality has been documented.
- [ ] Update [CHANGELOG.md](./../CHANGELOG.md)
- [ ] Commits are signed per the DCO using --signoff