@ohltyler 2023-07-21
> Overall, this looks great! Could you please include unit tests (UT) to ensure quality?

Added in latest commit - also enforces that the `hasReference` field is propagated as expected as well (from improvement in #4595) :)